useRichMenuDialog hookの作成
